svg/text/parsing/text-decoration-style-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>SVG Text CSS Decoration Test: getComputedStyle().textDecorationStyle</title><metadata><h:link ref="https://drafts.csswg.org/css-text-decor-3/#text-decoration-style-property" rel="help"/><h:meta content="SVG text-decoration-style computed value is as specified." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("text-decoration-style", "solid");
test_computed_value("text-decoration-style", "double");
test_computed_value("text-decoration-style", "dotted");
test_computed_value("text-decoration-style", "dashed");
test_computed_value("text-decoration-style", "wavy");

  ]]></script></svg>
